requirements.txt
astroid==2.3.0
Click==7.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.2
packaging==19.2
Pillow==6.1.0
pylint==2.4.1
pyparsing==2.4.2
redis==3.3.8
scikit-build==0.10.0
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
